+++ src/r_rational.rs
@@ -7,6 +7,7 @@ use crate::{
     exception,
     integer::Integer,
     into_value::IntoValue,
+    numeric::Numeric,
     ruby_handle::RubyHandle,
     try_convert::TryConvert,
     value::{private, NonZeroValue, ReprValue, Value},
@@ -144,6 +145,8 @@ unsafe impl private::ReprValue for RRational {
         Self(NonZeroValue::new_unchecked(val))
     }
 }
+
+impl Numeric for RRational {}
 
 impl ReprValue for RRational {}